+++ chakracore-cxx/lib/Runtime/Types/DynamicObject.h
@@ -3,6 +3,8 @@
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 //-------------------------------------------------------------------------------------------------------
 #pragma once
+#include "Common/VTableValue.h"
+
 class ScriptSite;
 namespace Js
 {
